+++ shared/include/srvros/syscall_numbers.h
@@ -102,6 +102,11 @@
 #define SYS_NET_PEEK 97
 #define SYS_PIPE_SEND_RIGHTS 98
 #define SYS_PIPE_RECV_RIGHTS 99
+#define SYS_UNIX_BIND 100
+#define SYS_UNIX_LISTEN 101
+#define SYS_UNIX_CONNECT 102
+#define SYS_UNIX_ACCEPT 103
+#define SYS_UNIX_UNLINK 104
 
 #define SRV_ABI_VERSION 1
 
@@ -186,6 +191,8 @@ struct srv_winsize {
 
 #define SRV_RIGHTS_MAX 4
 #define SRV_MSG_CTRUNC 0x08
+
+#define SRV_UNIX_PATH_MAX 108
 
 #define SRV_NETERR_NONE 0
 #define SRV_NETERR_RESET 104